Refactor IndentedTemplateComposer into modular architecture
Refactored the 920-line IndentedTemplateComposer service into smaller,
focused modules following Single Responsibility Principle:

**New Structure:**
- template_loader.py: Template loading and caching (~270 lines)
- code_generator.py: Code snippet generation (~220 lines)
- agent_template_composer.py: Agent template composition (~310 lines)
- function_template_composer.py: Function template composition (~220 lines)
- composer.py: Main orchestrator with delegation (~550 lines)
- __init__.py: Backwards-compatible exports

**Key Improvements:**
- Separated concerns: Each class has a single, well-defined responsibility
- Improved testability: Components can be tested independently
- Maintained 100% backwards compatibility: All existing tests pass
- Preserved public API: External code requires no changes
- Enhanced maintainability: Easier to understand and modify individual components

**Architecture:**
The refactored IndentedTemplateComposer acts as an orchestrator that
delegates to specialized components while maintaining all legacy methods
for backwards compatibility. Tests can still mock the main class methods,
which now properly delegate to the sub-components.

All 30 unit tests pass successfully, confirming backwards compatibility.

Fixes: Reduces code complexity and improves maintainability
Breaking Changes: None - 100% backwards compatible
